+++ network/index.js
@@ -1,6 +1,7 @@
 const currencyForNetwork = require('./currency_for_network');
 const executeProbe = require('./execute_probe');
 const getForwards = require('./get_forwards');
+const getIdentity = require('./get_identity');
 const getPeers = require('./get_peers');
 const multiPathPayment = require('./multi_path_payment');
 const multiPathProbe = require('./multi_path_probe');
@@ -20,6 +21,7 @@ module.exports = {
   currencyForNetwork,
   executeProbe,
   getForwards,
+  getIdentity,
   getPeers,
   multiPathPayment,
   multiPathProbe,